split up input data DTO into basic data and extended data used only with multienergy modelling
